refactor(website): split EmailsVisual into a folder structure
EmailsVisual.tsx shell + barrel, with ThreadRow in components/, the
EmailThread type in types/, and THREADS in data/. RecordTabHeader stays
shared at the product-feature level.
