+++ packages/twenty-website/src/sections/product-feature/EmailsVisual/index.ts
@@ -1,0 +1,1 @@
+export { EmailsVisual } from './EmailsVisual';
