+++ packages/twenty-website/src/sections/product-feature/EmailsVisual/types/email-thread.ts
@@ -1,0 +1,13 @@
+type Participant = {
+  avatarUrl: string;
+  name: string;
+};
+
+export type EmailThread = {
+  date: string;
+  messageCount: number;
+  participants: Participant[];
+  preview?: string;
+  shared: boolean;
+  subject?: string;
+};
